+++ javascript_client/src/subscriptions/__tests__/ActionCableLinkTest.ts
@@ -172,6 +172,31 @@ describe("ActionCableLink", () => {
     expect(subscription.params["test"]).toEqual(1)
   })
 
+  it("generates a unique channelId for each subscription", () => {
+    var link = new ActionCableLink(options)
+    var channelIds = new Set<string>()
+    var subscriptions: any[] = []
+
+    for (var i = 0; i < 1000; i++) {
+      var observable = link.request(operation, null as any)
+      var subscription: any = (observable.subscribe(() => null) as any)._cleanup
+      channelIds.add(subscription.params.channelId)
+      subscriptions.push(subscription)
+    }
+
+    expect(channelIds.size).toBe(1000)
+
+    subscriptions.forEach(function(s) { s.unsubscribe() })
+  })
+
+  it("accepts an injected channel ID function", () => {
+    var link = new ActionCableLink({...options, createChannelId: () => "Channel-ID" })
+    var observable = link.request(operation, null as any)
+    var subscription: any = (observable.subscribe(() => null) as any)._cleanup
+    expect(subscription.params.channelId).toEqual("Channel-ID")
+    subscription.unsubscribe()
+  })
+
   it('allows passing custom callbacks', () => {
     var connected = jest.fn()
     var received = jest.fn()