+++ javascript_client/src/subscriptions/createActionCableHandler.ts
@@ -1,4 +1,5 @@
 import type { Consumer } from "@rails/actioncable"
+import defaultChannelId from "./defaultChannelId"
 
 /**
  * Create a Relay Modern-compatible subscription handler.
@@ -12,12 +13,13 @@ interface ActionCableHandlerOptions {
   operations?: { getOperationId: Function}
   channelName?: string
   clientName?: string
+  createChannelId?: () => string
 }
 
 function createActionCableHandler(options: ActionCableHandlerOptions) {
+  const createChannelId = options.createChannelId || defaultChannelId
   return function (operation: { text: string, name: string, id?: string }, variables: object, _cacheConfig: object, observer: {onError: Function, onNext: Function, onCompleted: Function}) {
-    // unique-ish
-    var channelId = Math.round(Date.now() + Math.random() * 100000).toString(16)
+    var channelId = createChannelId()
     var cable = options.cable
     var operations = options.operations
     var subscribed = true